update devices -> providers in voice assembly / namespace, rename mr -> wmr in devices folder
