+++ Assets/MixedRealityToolkit.Providers/WindowsVoiceInput/MixedRealityToolkit.Providers.WindowsVoiceInput.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "Microsoft.MixedReality.Toolkit.Devices.WindowsVoiceInput",
+    "name": "Microsoft.MixedReality.Toolkit.Providers.WindowsVoiceInput",
     "references": [
         "Microsoft.MixedReality.Toolkit",
         "Microsoft.MixedReality.Toolkit.Core.Utilities.Async"
