--- Assets/MixedRealityToolkit/_Core/Devices/MixedReality/README.md
@@ -1,3 +1,0 @@
-# Mixed Reality Toolkit - Devices - Mixed Reality
-
-Device settings and definitions for the Microsoft Mixed Reality platform
