--- Assets/MixedRealityToolkit/_Core/Devices/MixedReality/README.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 62a6ccd44c244aba88f9176877447eba
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
