+++ Assets/MixedRealityToolkit/_Core/Devices/WindowsMixedReality.meta
