+++ Assets/MixedRealityToolkit/_Core/Devices/WindowsMixedReality/WindowsMixedRealityDeviceManager.cs
